requirements.txt
# requirements.txt

Flask
hdbcli
python-dotenv
Flask-SQLAlchemy
Flask-APScheduler
psutil